Fix auto admin user id creation.
The app was previously hard coded to make the admin id '1'.  It also
failed to initiate the mongo user id counter to 1.

The fix forces the admin account creation to use the getNextInt
function like all other users.  This also fixes the odd case of
removing the admin, recreating and being automatically logged in.