Fix app not starting when called directly. #2
The previous fix to make the tests wait for the app to start prevented
the app from starting when not using the tests.  This fix adds a check
to determine if the app.js is called using require or directly from
node and either waits or starts the app appropriately.